fix(core): repair nx reset
