+++ packages/nx/src/daemon/cache.ts
@@ -43,6 +43,23 @@ export async function safelyCleanUpExistingProcess(): Promise<void> {
   if (daemonProcessJson && daemonProcessJson.processId) {
     try {
       process.kill(daemonProcessJson.processId);
+      await new Promise<void>((resolve, reject) => {
+        let count = 0;
+        const interval = setInterval(() => {
+          try {
+            process.kill(daemonProcessJson.processId, 0);
+          } catch (e) {
+            clearInterval(interval);
+            resolve();
+          }
+          if ((count += 1) > 200) {
+            clearInterval(interval);
+            reject(
+              `Daemon process ${daemonProcessJson.processId} didn't exit after 2 seconds.`
+            );
+          }
+        }, 10);
+      });
     } catch {}
   }
   deleteDaemonJsonProcessCache();
