+++ packages/nx/src/project-graph/project-graph.ts
@@ -1,31 +1,32 @@
+import { performance } from 'perf_hooks';
+import { readNxJson } from '../config/nx-json';
+import { ProjectGraph } from '../config/project-graph';
 import {
-  readFileMapCache,
-  readProjectGraphCache,
-  writeCache,
-} from './nx-deps-cache';
+  ProjectConfiguration,
+  ProjectsConfigurations,
+} from '../config/workspace-json-project-json';
+import { daemonClient } from '../daemon/client/client';
+import { markDaemonAsDisabled, writeDaemonLogs } from '../daemon/tmp-dir';
+import { fileExists } from '../utils/fileutils';
+import { output } from '../utils/output';
+import { stripIndents } from '../utils/strip-indents';
+import { workspaceRoot } from '../utils/workspace-root';
 import {
   CreateDependenciesError,
   ProcessDependenciesError,
   ProcessProjectGraphError,
   buildProjectGraphUsingProjectFileMap,
 } from './build-project-graph';
-import { output } from '../utils/output';
-import { markDaemonAsDisabled, writeDaemonLogs } from '../daemon/tmp-dir';
-import { ProjectGraph } from '../config/project-graph';
-import { stripIndents } from '../utils/strip-indents';
 import {
-  ProjectConfiguration,
-  ProjectsConfigurations,
-} from '../config/workspace-json-project-json';
-import { daemonClient } from '../daemon/client/client';
-import { fileExists } from '../utils/fileutils';
-import { workspaceRoot } from '../utils/workspace-root';
-import { performance } from 'perf_hooks';
+  readFileMapCache,
+  readProjectGraphCache,
+  writeCache,
+} from './nx-deps-cache';
+
 import {
   retrieveProjectConfigurations,
   retrieveWorkspaceFiles,
 } from './utils/retrieve-workspace-files';
-import { readNxJson } from '../config/nx-json';
 import {
   ConfigurationResult,
   ConfigurationSourceMaps,
@@ -39,6 +40,7 @@ import {
 } from './error-types';
 import { DaemonProjectGraphError } from '../daemon/daemon-project-graph-error';
 import { loadNxPlugins, LoadedNxPlugin } from './plugins/internal-api';
+import { ProjectGraphError } from './project-graph-error';
 
 /**
  * Synchronously reads the latest cached copy of the workspace's ProjectGraph.
@@ -176,67 +178,6 @@ export async function buildProjectGraphAndSourceMapsWithoutDaemon() {
       writeCache(projectFileMapCache, projectGraph);
     }
     return { projectGraph, sourceMaps };
-  }
-}
-
-export class ProjectGraphError extends Error {
-  readonly #errors: Array<
-    | CreateNodesError
-    | MergeNodesError
-    | ProjectsWithNoNameError
-    | ProjectsWithConflictingNamesError
-    | ProcessDependenciesError
-    | ProcessProjectGraphError
-  >;
-  readonly #partialProjectGraph: ProjectGraph;
-  readonly #partialSourceMaps: ConfigurationSourceMaps;
-
-  constructor(
-    errors: Array<
-      | CreateNodesError
-      | MergeNodesError
-      | ProjectsWithNoNameError
-      | ProjectsWithConflictingNamesError
-      | ProcessDependenciesError
-      | ProcessProjectGraphError
-    >,
-    partialProjectGraph: ProjectGraph,
-    partialSourceMaps: ConfigurationSourceMaps
-  ) {
-    super(`Failed to process project graph.`);
-    this.name = this.constructor.name;
-    this.#errors = errors;
-    this.#partialProjectGraph = partialProjectGraph;
-    this.#partialSourceMaps = partialSourceMaps;
-    this.stack = `${this.message}\n  ${errors
-      .map((error) => error.stack.split('\n').join('\n  '))
-      .join('\n')}`;
-  }
-
-  /**
-   * The daemon cannot throw errors which contain methods as they are not serializable.
-   *
-   * This method creates a new {@link ProjectGraphError} from a {@link DaemonProjectGraphError} with the methods based on the same serialized data.
-   */
-  static fromDaemonProjectGraphError(e: DaemonProjectGraphError) {
-    return new ProjectGraphError(e.errors, e.projectGraph, e.sourceMaps);
-  }
-
-  /**
-   * This gets the partial project graph despite the errors which occured.
-   * This partial project graph may be missing nodes, properties of nodes, or dependencies.
-   * This is useful mostly for visualization/debugging. It should not be used for running tasks.
-   */
-  getPartialProjectGraph() {
-    return this.#partialProjectGraph;
-  }
-
-  getPartialSourcemaps() {
-    return this.#partialSourceMaps;
-  }
-
-  getErrors() {
-    return this.#errors;
   }
 }
